keepai/services/ai/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
langchain==0.1.0
langchain-google-genai==0.0.5
pydantic==2.5.2
requests==2.31.0
gunicorn==21.2.0 